The create() method 0)The render() method 1)The dispose() 2) Having fun with the debugger and Code Hot Swapping In this section, we are
Summary In this chapter, you learned a lot about LibGDX and how all the projects of an application work together. We covered LibGDX's backends, modules, and starter classes. Additionally, we covered what the application life cycle and corresponding interface are and how they are meant to work. The debugger has been used to inspect the demo application at runtime, and furthermore we made use of the  JVM Code Hot Swapping feature. We now know the basics of the LibGDX applications, so we are ready to start developing a real game. We will start at the very beginning of the development  cycle step by step. As LibGDX is a framework and not a game engine, we first  have to build our own engine. So, we will learn how to create an appropriate program architecture in the next chapter that is suitable to handle our gam