Putting it all together
We will now begin to fill in the stub methods with some life. The game loop is a good starting point; it is our driving engine that keeps the game world updated and rendered in a continuous way. After this, we will add some sprites and verify that the updating and rendering mechanism is working fine. In order to manipulate the world and game objects, controls are added to receive and react on user input. Finally, the CameraHelper class will be implemented to allow us to move around freely in the game world and to select a game object of our choice that the camera is supposed to follow.

------------------------------------
Building the game loop